refactor: :construction: updated side modal component moved to core
