+++ packages/core/package.json
@@ -36,6 +36,7 @@
     }
   ],
   "devDependencies": {
+    "@faststore/ui": "^1.5.3",
     "@babel/core": "^7.16.12",
     "@size-limit/preset-small-lib": "^7.0.5",
     "@storybook/addon-essentials": "^6.4.17",
@@ -43,13 +44,17 @@
     "@storybook/addon-links": "^6.4.17",
     "@storybook/addons": "^6.4.17",
     "@storybook/react": "^6.4.17",
+    "@storybook/preset-scss": "^1.0.3",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "babel-loader": "^8.2.3",
+    "css-loader": "5.2.7",
     "husky": "^7.0.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
+    "sass": "^1.49.7",
+    "sass-loader": "10.1.1",
     "size-limit": "^7.0.5",
     "tsdx": "^0.14.1",
     "tslib": "^2.3.1",
