--- packages/core/src/components/CoreTest/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const CoreTest = () => {
-  return <div>Core Test</div>;
-};
-
-export default CoreTest;
