+++ packages/core/src/components/NameInitials/index.tsx
