+++ packages/core/stories/components/MenuIcons/MenuIcons.scss
@@ -16,6 +16,10 @@
         color: #000000;
         margin-left: 10px;
         text-decoration: none;
+
+        &:hover {
+          text-decoration: underline;
+        }
       }
     }
   }
