+++ packages/core/stories/components/MenuIcons/MenuIcons.stories.tsx
