+++ packages/core/stories/components/NameInitials/NameInitials.scss
