+++ packages/core/stories/components/NameInitials/NameInitials.stories.tsx
