+++ packages/home/package.json
@@ -37,26 +37,20 @@
   ],
   "devDependencies": {
     "@babel/core": "^7.16.12",
-    "@faststore/ui": "^1.5.3",
     "@size-limit/preset-small-lib": "^7.0.5",
     "@storybook/addon-essentials": "^6.4.17",
     "@storybook/addon-info": "^5.3.21",
     "@storybook/addon-links": "^6.4.17",
     "@storybook/addons": "^6.4.17",
-    "@storybook/preset-scss": "^1.0.3",
     "@storybook/react": "^6.4.17",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "babel-loader": "^8.2.3",
-    "css-loader": "5.2.7",
     "husky": "^7.0.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^17.0.2",
-    "sass": "^1.49.7",
-    "sass-loader": "10.1.1",
     "size-limit": "^7.0.5",
-    "style-loader": "2.0.0",
     "tsdx": "^0.14.1",
     "tslib": "^2.3.1",
     "typescript": "^4.5.5"
