+++ packages/home/src/components/HomeTest/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const HomeTest = () => {
+  return <div>Home Test</div>;
+};
+
+export default HomeTest;
