+++ packages/home/src/index.ts
@@ -1,6 +1,1 @@
-import SideModal from './components/SideModal';
-import LoginButton from './components/Login';
-import NameInitials from './components/NameInitials';
-import MenuIcons from './components/MenuIcons';
-
-export { SideModal, LoginButton, NameInitials, MenuIcons };
+export * from './components';
