+++ packages/home/stories/Thing.stories.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Meta, Story } from '@storybook/react';
-import ExampleComponent, { Props } from '../src/ExampleComponent';
+import { Thing, Props } from '../src';
 
 const meta: Meta = {
   title: 'Welcome',
-  component: ExampleComponent,
+  component: Thing,
   argTypes: {
     children: {
       control: {
@@ -19,7 +19,7 @@ const meta: Meta = {
 
 export default meta;
 
-const Template: Story<Props> = (args) => <ExampleComponent {...args} />;
+const Template: Story<Props> = (args) => <Thing {...args} />;
 
 // By passing using the Args format for exported stories, you can control the props for a component for reuse in a test
 // https://storybook.js.org/docs/react/workflows/unit-testing
